Final Tweaks
Controls First Time backup. Asteroid and Comet spawning tweaked. Drill Rush buffed.

+++ Crashteroids/Assets/vanDriel_Joas/My Scripts/MyCometSpawn.cs
@@ -4,8 +4,8 @@ public class MyCometSpawn : MonoBehaviour
 {
     public AsteroidSpawnerAltered scriptbase;
     public GameObject cometPrefab;
-    public int spawnDelayLow;
-    public int spawnDelayHigh;
+    public float spawnDelayLow;
+    public float spawnDelayHigh;
     public float spawnChance;
     private bool spawnAllow;
     private float spawnTimer;
